Refactor: Split DD_ChangeGameWorker() into multiple tasks
Game change tasks are now processed by Con_BusyList(). Also improved
animation, by updating progress more frequently.

+++ doomsday/engine/portable/include/game.h
@@ -106,8 +106,7 @@ const ddstring_t* Game_BindingConfig(Game* game);
  * @param rclass  Class of resource to collect.
  * @return  Vector of selected resource records.
  */
-struct AbstractResource_s* const* Game_Resources(Game* game, resourceclass_t rclass,
-    size_t* count);
+struct AbstractResource_s* const* Game_Resources(Game* game, resourceclass_t rclass, int* count);
 
 /**
  * \note Unless caller is the resource locator then you probably shouldn't be calling.
